+++ src/app/app-routing.module.ts
@@ -1,7 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { ProductLayoutComponent } from './products/components/product-layout/product-layout.component';
+import { ProductListComponent } from './products/components/product-list/product-list.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    // The router will match this route if the URL requested
+    // doesn't match any paths for routes defined in our configuration
+    path: '**',
+    component: ProductListComponent
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],