+++ src/app/core/components/login/login.component.ts
@@ -1,0 +1,61 @@
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { AuthService } from '../../services/auth.service';
+
+@Component({
+  selector: 'app-login',
+  templateUrl: './login.component.html',
+  styleUrls: ['./login.component.css']
+})
+export class LoginComponent implements OnInit, OnDestroy {
+
+  message: string | undefined;
+
+  private unsubscribe: Subject<void> = new Subject();
+
+  constructor(
+    public authService: AuthService,
+    private router: Router) {
+  }
+
+  ngOnDestroy(): void {
+    this.unsubscribe.complete();
+  }
+
+  onLogin(): void {
+    this.message = 'Trying to log in ...';
+    const observer = {
+      next: () => {
+        this.setMessage();
+        if (this.authService.isLoggedIn) {
+          const redirect = this.authService.redirectUrl
+            ? this.authService.redirectUrl
+            : '/admin/products';
+          this.router.navigate([redirect]);
+        }
+      },
+      error: (err: any) => console.log(err),
+      complete: () => console.log('[takeUntil] complete')
+    };
+    this.authService
+      .login()
+      .pipe(takeUntil(this.unsubscribe))
+      .subscribe(observer);
+  }
+
+  onLogout(): void {
+    this.authService.logout();
+    this.setMessage();
+  }
+
+  private setMessage(): void {
+    this.message = 'Logged ' + (this.authService.isLoggedIn ? 'in' : 'out');
+  }
+
+  ngOnInit(): void {
+    this.setMessage();
+  }
+
+}