+++ src/app/core/core.module.ts
@@ -2,6 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FirstComponent } from './components/first/first.component';
 import { SharedModule } from '../shared/shared.module';
+import { CartRoutingModule } from './core-routing.module';
+import { LoginComponent } from './components/login/login.component';
 
 @NgModule({
   imports: [
@@ -9,8 +11,9 @@ import { SharedModule } from '../shared/shared.module';
     SharedModule
   ],
   exports: [
-    FirstComponent
+    FirstComponent,
+    CartRoutingModule
   ],
-  declarations: [FirstComponent]
+  declarations: [FirstComponent, LoginComponent]
 })
 export class CoreModule { }