+++ src/app/shared/pipes/order-by.pipe.spec.ts
@@ -5,7 +5,7 @@ import { OrderByPipe } from './order-by.pipe';
 
 describe('Pipe: OrderBye', () => {
   it('create an instance', () => {
-    let pipe = new OrderByPipe();
+    const pipe = new OrderByPipe();
     expect(pipe).toBeTruthy();
   });
 });